# Conflicts:
#	app-android/src/main/java/com/ustadmobile/port/android/impl/UstadApp.kt
#	app-ktor-server/src/main/kotlin/com/ustadmobile/lib/rest/UmRestApplication.kt
#	build.gradle
#	core/src/jvmTest/kotlin/com/ustadmobile/core/torrent/TestUstadTorrentManager.kt
#	core/src/jvmTest/kotlin/com/ustadmobile/core/util/UstadTestRule.kt
#	lib-database-android/schemas/com.ustadmobile.core.db.UmAppDatabase/88.json
#	lib-database/src/commonMain/kotlin/com/ustadmobile/core/db/UmAppDatabase.kt

+++ core/src/androidMain/kotlin/com/ustadmobile/core/catalog/contenttype/VideoTypePluginAndroid.kt
@@ -15,7 +15,7 @@ import com.ustadmobile.core.contentjob.*
 import com.ustadmobile.core.db.JobStatus
 import com.ustadmobile.core.db.UmAppDatabase
 import com.ustadmobile.core.io.ext.*
-import com.ustadmobile.core.torrent.UstadTorrentManager
+import com.ustadmobile.core.network.NetworkProgressListenerAdapter
 import com.ustadmobile.core.util.DiTag
 import com.ustadmobile.core.util.ext.*
 import com.ustadmobile.door.ext.toDoorUri
@@ -44,7 +44,7 @@ class VideoTypePluginAndroid(
         private var context: Any,
         private val endpoint: Endpoint,
         override val di: DI,
-        private val uploader: ContentPluginUploader = DefaultContentPluginUploader()
+        private val uploader: ContentPluginUploader = DefaultContentPluginUploader(di)
 ) : VideoTypePlugin() {
 
     private val VIDEO_ANDROID = "VideoPluginAndroid"
@@ -55,11 +55,7 @@ class VideoTypePluginAndroid(
 
     private val defaultContainerDir: File by di.on(endpoint).instance(tag = DiTag.TAG_DEFAULT_CONTAINER_DIR)
 
-    private val torrentDir: File by di.on(endpoint).instance(tag = DiTag.TAG_TORRENT_DIR)
-
     private val httpClient: HttpClient = di.direct.instance()
-
-    private val ustadTorrentManager: UstadTorrentManager by di.on(endpoint).instance()
 
     override suspend fun extractMetadata(uri: DoorUri, process: ContentJobProcessContext): MetadataResult? {
         return getEntry(uri, process)
@@ -74,10 +70,8 @@ class VideoTypePluginAndroid(
             val videoUri = DoorUri.parse(uri)
             val contentNeedUpload = !videoUri.isRemote()
             val localUri = process.getLocalOrCachedUri()
-            val trackerUrl = db.siteDao.getSiteAsync()?.torrentAnnounceUrl
-                    ?: throw IllegalArgumentException("missing tracker url")
-
-            val progressSize = if(contentNeedUpload) 2 else 1
+            contentJobItem.updateTotalFromLocalUriIfNeeded(localUri, contentNeedUpload,
+                jobProgress, context, di)
 
             val videoTempDir = makeTempDir(prefix = "tmp")
             val newVideo = File(videoTempDir,
@@ -141,13 +135,21 @@ class VideoTypePluginAndroid(
 
                                 override fun onProgress(id: String, progress: Float) {
                                     Napier.d(tag = VIDEO_ANDROID, message = "progress at value ${progress * 100}")
-                                    contentJobItem.cjiItemProgress = ((progress * contentJobItem.cjiItemTotal) / progressSize).toLong()
+                                    var cjiProgress = (progress * contentJobItem.cjiItemTotal)
+                                    if(contentNeedUpload)
+                                        cjiProgress /= 2.toFloat()
+
+                                    contentJobItem.cjiItemProgress = cjiProgress.toLong()
                                     jobProgress.onProgress(contentJobItem)
                                 }
 
                                 override fun onCompleted(id: String, trackTransformationInfos: MutableList<TrackTransformationInfo>?) {
                                     Napier.d(tag = VIDEO_ANDROID, message = "completed transform")
-                                    contentJobItem.cjiItemProgress = contentJobItem.cjiItemTotal / progressSize
+                                    var cjiProgress = contentJobItem.cjiItemTotal
+                                    if(contentNeedUpload)
+                                        cjiProgress /= 2
+
+                                    contentJobItem.cjiItemProgress = cjiProgress
                                     jobProgress.onProgress(contentJobItem)
                                     videoCompleted.complete(true)
                                 }
@@ -200,24 +202,14 @@ class VideoTypePluginAndroid(
                         repo.addContainerFromUri(container.containerUid, localUri, context, di,
                                 localUri.getFileName(context),
                                 ContainerAddOptions(containerFolderUri))
-
-                        contentJobItem.cjiItemProgress = contentJobItem.cjiItemTotal / progressSize
-                        jobProgress.onProgress(contentJobItem)
                     }
                     videoTempDir.delete()
 
-                    repo.writeContainerTorrentFile(
-                            container.containerUid,
-                            DoorUri.parse(torrentDir.toURI().toString()),
-                            trackerUrl, containerFolderUri
-                    )
-
-                    val containerUidFolder = File(containerFolderUri.toFile(), container.containerUid.toString())
-                    containerUidFolder.mkdirs()
-                    ustadTorrentManager.addTorrent(container.containerUid, containerUidFolder.path)
-
                     contentJobItem.cjiContainerUid = container.containerUid
-                    db.contentJobItemDao.updateContentJobItemContainer(contentJobItem.cjiUid, container.containerUid)
+                    db.contentJobItemDao.updateContentJobItemContainer(contentJobItem.cjiUid,
+                        container.containerUid)
+                    contentJobItem.updateTotalFromContainerSize(contentNeedUpload, db,
+                        jobProgress)
 
                     contentJobItem.cjiConnectivityNeeded = true
                     db.contentJobItemDao.updateConnectivityNeeded(contentJobItem.cjiUid, true)
@@ -230,11 +222,10 @@ class VideoTypePluginAndroid(
 
                 }
 
-                val torrentFileBytes = File(torrentDir, "${contentJobItem.cjiContainerUid}.torrent").readBytes()
                 if(contentNeedUpload) {
-                    uploader.upload(
-                            contentJobItem, jobProgress, httpClient, torrentFileBytes,
-                            endpoint)
+                    uploader.upload(contentJobItem,
+                        NetworkProgressListenerAdapter(jobProgress, contentJobItem),
+                        httpClient, endpoint)
                 }
 
                 return@withContext ProcessResult(JobStatus.COMPLETE)
@@ -244,7 +235,9 @@ class VideoTypePluginAndroid(
                     mediaTransformer.release()
                     newVideo.delete()
                     videoTempDir.deleteRecursively()
-                    localUri.toFile().delete()
+                    if(videoUri.isRemote()){
+                        localUri.toFile().delete()
+                    }
                 }
 
                 throw c
@@ -277,12 +270,6 @@ class VideoTypePluginAndroid(
             }
             MetadataResult(entry, pluginId)
         }
-    }
-
-    companion object {
-
-        val MEDIA_TYPE_TORRENT = "application/x-bittorrent"
-
     }
 
 }
